+++ frontend/src/update_address_flow.tsx
@@ -13,11 +13,12 @@ import {
   Button,
   MainContent,
   Callout,
+  H4,
 } from '@votingworks/ui';
 import { useState } from 'react';
 import { Column, Row } from './layout';
 import { NoNavScreen } from './nav_screen';
-import { VoterName } from './shared_components';
+import { TitledCard, VoterName } from './shared_components';
 import { AddressInputGroup } from './address_input_group';
 import { changeVoterAddress } from './api';
 
@@ -42,9 +43,11 @@ function createBlankAddress(): VoterAddressChangeRequest {
 }
 
 function UpdateAddressScreen({
+  voter,
   onConfirm,
   onCancel,
 }: {
+  voter: Voter;
   onConfirm: (address: VoterAddressChangeRequest) => void;
   onCancel: () => void;
 }): JSX.Element {
@@ -60,7 +63,17 @@ function UpdateAddressScreen({
       </MainHeader>
       <MainContent>
         <Column style={{ gap: '1rem' }}>
-          <AddressInputGroup address={address} onChange={setAddress} />
+          <TitledCard
+            title={
+              <H4>
+                <VoterName voter={voter} />
+              </H4>
+            }
+          >
+            <Column style={{ gap: '1rem' }}>
+              <AddressInputGroup address={address} onChange={setAddress} />
+            </Column>
+          </TitledCard>
           {address.streetNumber.trim() !== '' &&
             address.streetName !== '' &&
             !isAddressValid && (
@@ -116,6 +129,7 @@ export function UpdateAddressFlow({
     case 'update':
       return (
         <UpdateAddressScreen
+          voter={voter}
           onConfirm={(addressChangeData) => {
             setFlowState({ step: 'printing' });
             changeVoterAddressMutation.mutate(
